pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "chess-analyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "click>=8.3.1",
    "litellm>=1.40.0",
    "instructor>=1.3.3",
    "pydantic>=2.12.5",
    "python-chess>=1.999",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "rich>=14.3.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[project.scripts]
chess-analyst = "source.main:cli"

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["source*"]
